requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu124
absl-py==2.3.1
accelerate==1.10.1
antlr4-python3-runtime==4.8
audiolm-pytorch==2.0.6
audioread==3.1.0
av==13.1.0
beartype==0.16.1
bitarray==3.8.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Cython==3.2.1
decorator==5.2.1
docopt==0.6.2
editdistance==0.8.1
einops==0.8.1
einx==0.3.0
ema-pytorch==0.7.7
encodec==0.1.1
eval_type_backport==0.2.2
fairseq==0.12.2
filelock==3.19.1
fonttools==4.60.1
frozendict==2.4.7
fsspec==2025.10.0
Gammatone @ https://github.com/detly/gammatone/archive/master.zip
gateloop-transformer==0.2.5
gitdb==4.0.12
GitPython==3.1.45
grpcio==1.76.0
hf-xet==1.2.0
huggingface-hub==0.36.0
hydra-core==1.0.7
hyper-connections==0.2.1
HyperPyYAML==1.2.2
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
indic-num2words==1.3.2
inflect==7.5.0
Jinja2==3.1.6
jiwer==4.0.0
joblib==1.5.2
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.11.0
lightning-utilities==0.15.2
lion-pytorch==0.2.3
llvmlite==0.43.0
local-attention==1.11.2
lxml==6.0.2
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
masked_prosody_model==0.3.0
matplotlib==3.9.4
mdurl==0.1.2
mock==5.2.0
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
naturalspeech2-pytorch==0.1.8
networkx==3.2.1
nnAudio==0.3.1
nose==1.3.7
num2words==0.5.14
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.2.65
nvidia-cuda-cupti-cu12==12.4.99
nvidia-cuda-nvrtc-cu12==12.4.99
nvidia-cuda-runtime-cu12==12.4.99
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.0.44
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.5.119
nvidia-cusolver-cu12==11.6.0.99
nvidia-cusparse-cu12==12.3.0.142
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.4.99
omegaconf==2.0.6
packaging==25.0
pandas==2.3.3
pillow==11.3.0
platformdirs==4.4.0
pooch==1.8.2
portalocker==3.2.0
praat-parselmouth==0.4.7
protobuf==6.33.0
psutil==7.1.3
pycparser==2.23
pydantic==1.10.24
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytorch-warmup==0.2.0
pytz==2025.2
pyworld==0.3.5
pyworld-prebuilt==0.3.5.post2
PyYAML==6.0.3
RapidFuzz==3.13.0
regex==2025.11.3
requests==2.32.5
Resemblyzer==0.1.4
rich==14.2.0
rotary-embedding-torch==0.8.9
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
sacrebleu==2.5.1
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
sentencepiece==0.2.1
sentry-sdk==2.44.0
six==1.17.0
smmap==5.0.2
soundfile==0.13.1
soxr==1.0.0
speechbrain @ git+https://github.com/speechbrain/speechbrain.git@b7afd0350a9f6acaf8aa6dcd7dcffec25fa82e30
speechtokenizer @ file:///scratch/anon/anon/workspace/anonymization/ns2vc_private/SpeechTokenizer
SRMRpy @ git+https://github.com/minixc/srmrpy@e504686ea0e375165f13b34c9a531a3ffef9a4f9
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tokenizers==0.21.4
torchaudio==2.4.0+cu124
torchmetrics==1.8.2
torchvision==0.19.0+cu124
tqdm==4.67.1
transformers==4.49.0
triton==3.0.0
typeguard==4.4.4
typing==3.7.4.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vector-quantize-pytorch==1.25.2
vocos==0.1.0
wandb==0.23.0
webrtcvad==2.0.10
Werkzeug==3.1.3
zipp==3.23.0
